backend/requirements.txt
apturl==0.5.2
attrs==23.1.0
bcrypt==3.2.0
blinker==1.8.2
Brlapi==0.8.3
# Editable install with no version control (c-formatter-42==0.2.7)
-e /home/gouzi/.local/lib/python3.10/site-packages
certifi==2023.11.17
chardet==4.0.0
chrome-gnome-shell==0.0.0
click==8.1.7
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
duplicity==0.8.21
exceptiongroup==1.2.0
fasteners==0.14.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
future==0.18.2
guake==3.8.5
h11==0.14.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.4
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==3.0.2
monotonic==1.6
more-itertools==8.10.0
mysql-connector-python==9.1.0
netifaces==0.11.0
norminette==3.3.54
oauthlib==3.2.0
olefile==0.46
outcome==1.3.0.post0
paramiko==2.9.3
pbr==5.8.0
pexpect==4.8.0
Pillow==10.1.0
protobuf==3.12.4
ptyprocess==0.7.0
pycairo==1.20.1
pycodestyle==2.8.0
pycryptodomex==3.11.0
pycups==2.0.1
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
PySocks==1.7.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
pytz==2022.1
pyxattr==0.7.2
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
SecretStorage==3.3.1
selenium==4.15.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
systemd-python==234
tk==0.1.0
trio==0.23.1
trio-websocket==0.11.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
vboxapi==1.0
wadllib==1.3.6
Werkzeug==3.1.0
wsproto==1.2.0
xdg==5
xkit==0.0.0
youtube-dl==2021.12.17
zipp==1.0.0
